Device and driver instantiation with address
The way how the drivers are instantiated has been changed to take into account the fact that we now a need to provided the device address in the init function of the 'Device' class. These address are stored in a file 'devices_index.txt' located in the folder 'local_config' located beside the usit folder.